chore: move workshopSpeaker to speakers folder
